Add paging bar to errors report
When viewing the errors report - for large institutions you can have a
long list of errors - when typically you're only interested in recent
ones.

This change adds a paging bar to the top and bottom of the report and
has it display 100 records per page.

The underlying record lookup is also adjusted to only load the 100
records related to the page - instead of ALL of them, reducing the
memory footprint considerably.