Wrap colorbox in a click handler for EULA link
When calling colorbox directly on the EULA link, it doesn't always open
in an iframe if the student has previously declined the EULA - often it
will load the form inline, resulting in the browser being redirected to
the EULA on a blank page on the TII site instead of displaying on top of
the Moodle page.  This means that the user gets "stuck" on the TII site -
they're not returned to Moodle after either accepting or declining.